+++ service-proxy/src/main/java/no/mnemonic/services/common/api/proxy/client/ServiceMessageConverter.java
@@ -3,6 +3,7 @@ package no.mnemonic.services.common.api.proxy.client;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
 import lombok.CustomLog;
+import lombok.Getter;
 import lombok.NonNull;
 import no.mnemonic.services.common.api.ServiceContext;
 import no.mnemonic.services.common.api.proxy.Utils;
@@ -13,6 +14,7 @@ import no.mnemonic.services.common.api.proxy.serializer.Serializer;
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.Method;
+import java.util.Map;
 import java.util.UUID;
 
 import static no.mnemonic.services.common.api.proxy.Utils.fromTypes;
@@ -55,7 +57,7 @@ public class ServiceMessageConverter {
    * @return the deserialized response message
    * @throws Exception the deserialized exception, if the response contained an exception
    */
-  public Object readResponseMessage(InputStream response) throws Exception {
+  public Response readResponseMessage(InputStream response) throws Exception {
     ServiceResponseMessage responseMessage = mapper.readValue(response, ServiceResponseMessage.class);
     if (responseMessage.getException() != null) {
       if (LOGGER.isDebug()) LOGGER.debug(">> received exception [callID=%s]", responseMessage.getRequestID());
@@ -65,8 +67,19 @@ public class ServiceMessageConverter {
     if (LOGGER.isDebug()) {
       LOGGER.debug(">> received response [callID=%s]", responseMessage.getRequestID());
     }
-    if (responseMessage.getResponse() == null) return null;
-    return serializer.deserializeB64(responseMessage.getResponse());
+    if (responseMessage.getResponse() == null)  {
+      return new Response(null, responseMessage.getMetaData());
+    } else {
+      Object obj = serializer.deserializeB64(responseMessage.getResponse());
+      return new Response(obj, responseMessage.getMetaData());
+    }
+  }
+
+  @AllArgsConstructor
+  @Getter
+  public static class Response {
+    private final Object response;
+    private final Map<String, String> metaData;
   }
 
   //private methods